requirements.txt
Django==2.2.7
django-extensions==2.2.5
django-rest-framework==0.1.0
djangorestframework==3.10.3
numpy==1.17.3
pandas==0.25.3
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
